Shift location of model tests appropriately.
